requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.4
django-cleanup==4.0.0
django-crispy-forms==1.9.0
django-extensions==3.0.4
django-phonenumber-field==4.0.0
django-unixtimestampfield==0.4.0
idna==2.8
IMDbPY==6.8
lxml==4.5.2
memorised==1.1.0
numpy==1.20.3
pandas==1.2.4
phonenumbers==8.12.0
Pillow==7.2.0
#psutil==5.7.2
#psycopg2-binary==2.9.1
python-dateutil==2.8.1
python-memcached==1.59
python3-memcached==1.51
pytz==2018.9
requests==2.21.0
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
urllib3==1.24.3
